requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
attrs==23.2.0
Authlib==1.3.0
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
coloredlogs==15.0.1
configobj==5.0.8
configparser==6.0.1
contourpy==1.0.2
cryptography==42.0.2
cycler==0.12.1
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
Deprecated==1.2.14
distro==1.9.0
effdet==0.4.1
emoji==2.10.1
etelemetry==0.3.1
filelock==3.13.1
filetype==1.2.0
fitz==0.0.1.dev2
flatbuffers==23.5.26
fonttools==4.49.0
frontend==0.0.3
frozenlist==1.4.1
fsspec==2024.2.0
gradio==4.28.3
grpcio==1.62.1
grpcio-health-checking==1.62.1
grpcio-tools==1.62.1
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
importlib-resources==6.1.1
iopath==0.1.10
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.45
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langdetect==1.0.9
langsmith==0.1.27
layoutparser==0.3.4
looseversion==1.3.0
lxml==5.1.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.7.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.1
nibabel==5.2.1
nipype==1.8.6
nltk==3.8.1
numpy==1.24.4
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==0.28.0
opencv-python==4.8.0.76
orjson==3.9.15
packaging==23.2
pandas==2.0.3
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.4
pikepdf==8.11.0
pillow==10.2.0
pillow_heif==0.15.0
portalocker==2.8.2
protobuf==4.23.4
prov==2.0.0
pyarrow==15.0.1
pycocotools==2.0.7
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydot==2.0.0
pyparsing==3.0.9
pypdf==4.0.1
PyPDF2==3.0.1
pypdfium2==4.27.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-iso639==2024.2.7
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rapidfuzz==3.6.1
rdflib==7.0.0
regex==2023.12.25
requests==2.31.0
safetensors==0.3.2
scikit-learn==1.3.2
scipy==1.10.1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
starlette==0.37.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
timm==0.9.12
tokenizers==0.15.2
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.2
traits==6.3.2
transformers==4.37.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
unstructured==0.11.8
unstructured-client==0.18.0
unstructured-inference==0.7.23
unstructured.pytesseract==0.3.12
urllib3==1.26.18
uvicorn==0.28.1
validators==0.22.0
weaviate-client==4.5.4
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
